+++ examples/selectjsonblob.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2015, 2019, Oracle and/or its affiliates. All rights reserved. */
+/* Copyright (c) 2015, 2020, Oracle and/or its affiliates. All rights reserved. */
 
 /******************************************************************************
  *
@@ -19,12 +19,13 @@
  *   selectjsonblob.js
  *
  * DESCRIPTION
- *   Executes sample insert and query statements using a JSON column with BLOB storage.
+ *   Shows how to use a BLOB as a JSON column store.
  *
- *   Requires Oracle Database 12.1.0.2, which has extensive JSON datatype support.
+ *   Note: with Oracle Database 21c using the new JSON type is recommended
+ *   instead, see selectjson.js
+ *
+ *   Requires Oracle Database 12.1.0.2 or later.
  *   See https://www.oracle.com/pls/topic/lookup?ctx=dblatest&id=ADJSN
- *
- *   This example requires node-oracledb 1.13 or later.
  *
  *   This example uses Node 8's async/await syntax.
  *
@@ -32,6 +33,8 @@
 
 const oracledb = require('oracledb');
 const dbConfig = require('./dbconfig.js');
+
+oracledb.extendedMetaData = true;
 
 async function run() {
 
@@ -42,58 +45,81 @@ async function run() {
     connection = await oracledb.getConnection(dbConfig);
 
     if (connection.oracleServerVersion < 1201000200) {
-      throw new Error('This example only works with Oracle Database 12.1.0.2 or greater');
+      throw new Error('Using JSON requires Oracle Database 12.1.0.2 or later');
     }
 
-    const stmts = [
-      `DROP TABLE no_purchaseorder_b`,
+    console.log('1. Creating Table');
 
-      `CREATE TABLE no_purchaseorder_b (po_document BLOB CHECK (po_document IS JSON)) LOB (po_document) STORE AS (CACHE)`
-    ];
-
-    for (const s of stmts) {
-      try {
-        await connection.execute(s);
-      } catch(e) {
-        if (e.errorNum != 942)
-          console.error(e);
-      }
+    try {
+      await connection.execute(`DROP TABLE no_purchaseorder_b`);
+    } catch(e) {
+      if (e.errorNum != 942)
+        console.error(e);
     }
 
-    let result;
-
-    console.log('Inserting Data');
-    const data = { "userId": 2, "userName": "Bob", "location": "USA" };
-    const s = JSON.stringify(data);
-    const b = Buffer.from(s, 'utf8');
     await connection.execute(
-      `INSERT INTO no_purchaseorder_b (po_document) VALUES (:lobbv)`,
-      { lobbv: b }
-    );
+      `CREATE TABLE no_purchaseorder_b
+           (po_document BLOB CHECK (po_document IS JSON)) LOB (po_document) STORE AS (CACHE)`);
 
-    console.log('Selecting JSON stored in a BLOB column:');
+    console.log('2. Inserting Data');
+
+    const inssql = `INSERT INTO no_purchaseorder_b (po_document) VALUES (:bv)`;
+    const data = { "userId": 1, "userName": "Anna", "location": "Australia" };
+
+    if (oracledb.oracleClientVersion >= 2100000000 && connection.oracleServerVersion >= 2100000000 ) {
+      // Take advantage of direct binding of JavaScript objects
+      await connection.execute(inssql, { bv: { val: data, type: oracledb.DB_TYPE_JSON } });
+    } else {
+      // Insert the data as a JSON string
+      const s = JSON.stringify(data);
+      const b = Buffer.from(s, 'utf8');
+      await connection.execute(inssql, { bv: { val: b } });
+    }
+
+    console.log('3. Selecting JSON stored in a BLOB column');
+
+    let result, j;
+
     result = await connection.execute(
       `SELECT po_document
        FROM no_purchaseorder_b
-       WHERE JSON_EXISTS (po_document, '$.location')`,
-      [],
-      { fetchInfo: { "PO_DOCUMENT": { type: oracledb.BUFFER } } }  // Fetch as a Buffer instead of a Stream
+       WHERE JSON_EXISTS (po_document, '$.location')
+       OFFSET 0 ROWS FETCH NEXT 1 ROWS ONLY`
     );
-    if (result.rows.length === 0)
-      throw new Error('No results');
-    console.log(result.rows[0][0].toString('utf8'));
+    const d = await result.rows[0][0].getData();
+    j = await JSON.parse(d);
+    console.log('Query results: ', j);
 
-    console.log('Selecting a JSON value using "dotted" notation:');
-    if (connection.oracleServerVersion < 1202000000) {
-      throw new Error('This example only works with Oracle Database 12.2 or greater');
-    }
+    console.log('4. Using JSON_VALUE to extract a value from a JSON column');
+
     result = await connection.execute(
-      `SELECT pob.po_document.location
-       FROM no_purchaseorder_b pob`
+      `SELECT JSON_VALUE(po_document, '$.location')
+       FROM no_purchaseorder_b
+       OFFSET 0 ROWS FETCH NEXT 1 ROWS ONLY`
     );
-    if (result.rows.length === 0)
-      throw new Error('No results');
-    console.log(result.rows[0][0]);
+    console.log('Query results: ', result.rows[0][0]);  // just show first record
+
+    if (connection.oracleServerVersion >= 1202000000) {
+
+      console.log('5. Using dot-notation to extract a value from a JSON column');
+
+      result = await connection.execute(
+        `SELECT po.po_document.location
+         FROM no_purchaseorder_b po
+         OFFSET 0 ROWS FETCH NEXT 1 ROWS ONLY`
+      );
+      console.log('Query results: ', result.rows[0][0]);
+
+      console.log('6. Using JSON_OBJECT to extract relational data as JSON');
+
+      result = await connection.execute(
+        `SELECT JSON_OBJECT('key' IS d.dummy) dummy
+         FROM dual d`
+      );
+      for (let row of result.rows) {
+        console.log('Query results: ', row[0]);
+      }
+    }
 
   } catch (err) {
     console.error(err);