+++ src/njsErrors.c
@@ -84,6 +84,8 @@ static const char *njsErrorMessages[] = {
     "NJS-075: only one of connectString and connectionString can be used", // errDblConnectionString
     "NJS-076: connection request rejected. Pool queue length queueMax %d reached", // errQueueMax
     "NJS-077: Oracle Client library has already been initialized", // errClientLibAlreadyInitialized
+    "NJS-078: unsupported data type %u in JSON value", // errUnsupportedDataTypeInJson
+    "NJS-079: cannot convert from JavaScript value to JSON value", // errConvertToJsonValue
 };