+++ test/fetchAs.js
@@ -58,13 +58,19 @@ describe('56. fetchAs.js', function() {
     (oracledb.fetchAsString).should.eql([]);
 
     oracledb.fetchAsString=[oracledb.DATE];
-    (oracledb.fetchAsString).should.eql( [oracledb.DATE] );
+    should.deepEqual(oracledb.fetchAsString, [ oracledb.DATE ]);
 
     oracledb.fetchAsString = [ oracledb.NUMBER ];
-    (oracledb.fetchAsString).should.eql( [oracledb.NUMBER] );
+    should.deepEqual(oracledb.fetchAsString, [ oracledb.NUMBER ]);
 
     oracledb.fetchAsString = [ oracledb.DATE, oracledb.NUMBER ];
-    (oracledb.fetchAsString).should.eql( [oracledb.DATE, oracledb.NUMBER] );
+    should.deepEqual(oracledb.fetchAsString, [ oracledb.DATE, oracledb.NUMBER ]);
+
+    oracledb.fetchAsString = [ oracledb.DB_TYPE_JSON ];
+    should.deepEqual(oracledb.fetchAsString, [ oracledb.DB_TYPE_JSON ]);
+
+    oracledb.fetchAsString = [ oracledb.DB_TYPE_JSON, oracledb.DATE, oracledb.NUMBER ];
+    should.deepEqual(oracledb.fetchAsString, [ oracledb.DB_TYPE_JSON, oracledb.DATE, oracledb.NUMBER ]);
   });
 
   it('56.2 Fetch DATE column values as STRING - by-Column name', function(done) {